more re-export organization.
Trying out qualifying `Range` as a test. The idea is that relatively immovable
objects, like std::, can be qualified rather than used, so that the use section
looks a little less daunting for newcomers. Of course, this makes the code
itself more daunting because there's a lot more repetition. It's a tradeoff.