A bunch of changes
All I was trying to do was pass some drag-and-drop info up to the app, but I
found that I needed to repackage an Action message again and again to add
another bit of context that only the next level up had. For example, Dca noticed
that someone had dropped a Controller source on gain or pan, so it would have
been happy to emit a link_control, but it needed to know its param offsets
relative to the synth that contained it (gain = 0, pan = 1, but in Welsh I think
it's 32 + 0 and 32 + 1). So it couldn't do it itself. So it passed it up to the
synth, but then the synth needed to know its track ID, because ControlRouter is
local to a track. And so on. It was hard to add new stuff.

So I thought that Orchestrator could maintain a mapping of entity ID to track
ID, which would mean nobody had to remember TrackUid anymore. But then that
meant that I'd need to ditch my current approach of letting callers reach into
Orchestrator to get a Track, then into Track to get a ControlRouter etc.,
because Orchestrator would need to know more about how things mutated so that it
could keep its index up to date.

At this point I realized that it was time for a formal Orchestrator API, and the
Orchestrates trait was born. It will provide all the arrangement/composition
functionality at the top level (widgets can still do internal magic if they
want), and avoid exposing things like Track (only TrackUid) or ControlRouter.

But I didn't want to make the API just a flattened version of everything
reachable from Orchestrator, particularly Sequencer, ControlAtlas, which I know
will be changing a lot over time (and are very modular). So Tracks no longer
have their own internal ones of those; they take them just like other Entities.
This makes integration tests harder to write, because we have to provide things
we used to get for free.

I will need to develop Sequencer and Control traits, because those are
complicated to render in the timeline. TODO: adapt timeline widget to work with
these traits.

I also changed DragDropManager to use a crossbeam channel rather than queuing up
its own Events. That makes more sense for a long-lived, widely used thing like
DDM, versus an ephemeral widget that needs to conform to the Displays trait.