Sequencer - redesigned for composition
Removed the `SequencesNotes: Sequences` trait bound, which I could see was
moving in a horrible direction if each higher layer had to acknowledge the
behavior of the layer(s) below it. Instead, we now define SequencesMidi as the
bedrock that knows only about MIDI events, and then let Sequences<whatever>
compose behavior out of things below it. So Sequences<Note> contains a
SequencesMidi (but you don't know that), and Sequences<Pattern> contains a
Sequences<Note>. And so on.