Cleaning up Action cruft
There are many different ways I'm communicating what happened in the UI. It was
OK to experiment with them, but now it's time to clean up.

Way #1: just do it when it happens. This is preferred, but only if the widget
code was already single-purpose.

Way #2: Pass back the action in the widget. This would cause problems for
structs that implement Displays but don't have public widgets, because we'd need
a way for the struct to communicate an action.

Way #3: Implement Acts in the struct being viewed. The realization that allowed
me to get here vs #2 was that a struct can be rendered by multiple widgets
(e.g., TrackWidget and SignalChainWidget take care of two different major parts
of Track), but that doesn't mean that each widget needs its own WhateverAction
enum. Instead, they can simply share the struct's enum. Even if they can't both
produce the same actions, it doesn't feel bad to merge two sets of enums into
one.

I've gone through and cleaned up so that everything follows Way 1 or Way 3.
There will be a followup commit that moves at least one widget (Dca) to its
owner.