Patched up PianoRoll's crazy scrolling effect
egui is creepy. If you use ui.available_something(), there's a risk that your
content will increase the available width or height, which means that on the
next frame, you'll get a larger number, which leads to this odd creeping effect
that looks like a bad UI animation. At least I think that's why it's happening.

Generally I need to understand the zen of egui layout. I know it's not simple,
but I'm sure I can develop a working knowledge of it. In particular, I want to
know who should decide how big an area should be. Is it the parent, which
subtracts space from what it received? Or is it the child, which lays out what
it needs, and then demands that much space from the parent?