Integrate time domain widget
This is doing too much work on the UI thread; it's servicing a new crossbeam
channel that receives 64-sample buffers when Orchestrator produces them. If it
becomes an issue, we can spin up a thread that watches that channel and fills
the CircularSampleBuffer asynchronously.

We are also doing a StereoSample -> Sample conversion in the audio pipeline...
and then another copy of that whole buffer to get it into the message.