Streamlined Controls::work()
There are two modes.

One is you called me, so you know my uid. This was the None case in the old
work() method. It was far and away the most common case, so I removed the
argument.

Two is I own a bunch of Entities, so you don't know my uid. This was the Some()
case. I separated that out into a new trait, ControlsAsProxy, that restores the
original non-Option Uid argument.

Also tried moving Uid out of core, but that didn't quite work. Some motion but
not clean.