test entities are now a feature
Successful shower thought. I have been bedeviled by test entities
(TestInstrument et al.), because I want to use them all over the place, but I
didn't want them part of the official crate API. I wanted to stick them inside
`pub mod tests`, which is itself behind `#[cfg(test)]`, but crates don't seem to
support anything like "make this public but only for consumers that happen to
have declared the same `cfg` as the one we used," which makes sense when I write
it out like that. So I was stuck -- either restrict them to being used in the
crate where they live, or declare them as part of the crate exports.

The breakthrough was realizing that "test" had nothing to do with it; what I
want is a feature. If `dev-dependencies` lets you declare features for just the
tests and examples, then we're all set. I did this, and it seems to work great.

I decided to make `ensnare-entities-toy` a separate crate, even though to me
they'd be fine as another feature, because I thought it would be useful to show
how to create a crate of new instruments (and to keep myself from accidentally
breaking that ability).

Other changes

- Cleaned up *all* warnings
- Added a simple license checker into `precheck`

TODO

- separate the system entities from the ones that end users see. PianoRoll and
  ControlTrip are special, for example.