ControlTrip entity has access to ControlRouter
As hoped, we were easily able to give ControlTrip an Arc<RwLock<ControlRouter>>
at creation time. I still don't know how to present this info/interaction, but
the data flow is working.

There is a deadlock concern. If the trip widget causes an immediate
ControlRouter mutation, then it'll block forever because we already had a read()
on ControlRouter to get the list of links. I can avoid this by cloning the list
and releasing the read(), or by deferring the mutation Elm-style.