Fix spin and clean up old code
This was a cool bug. I was getting too clever. I wanted to address the services
only by channel, never directly, so I wasn't even saving any references to the
services after cloning their channels. So they were getting dropped! And I was
seeing channel recv() failures, which was causing receive loops to spin.

The AudioService is due for a cleanup. It seems to have a service inside a
service, but one ought to be just fine.