Bugfix: extents must take position arg into account
Got rid of Pattern's overloaded Add, which I realized was ambiguous: adding C to
a range could shift both start and end, or it could increase just the end. I
think the more correct interpretation is shifting (since you'd add a range to a
range, not a time to a range, to make it longer), but I second-guessed myself
after re-reading old code, so it's ambiguous. QED.