External MIDI routing is now a setting
Whew, this was harder than I thought it would be.

App owns settings and service event aggregator. Settings owns MIDI settings.
MIDI settings knows when the setting changes. It has no way of talking to its
owner, and it felt wrong to set up a whole crossbeam_channel apparatus for that.

I realized the Settings widget was the only place the setting would change, and
that it had access to both settings and MIDI settings. So when the setting
changes, it asks settings (via its new channel) to tell the app that the setting
changed. BUT we actually want the aggregator to know that, because it's the one
doing the MIDI routing from project to MIDI interface service.

Fortunately, once again, this proved simpler than getting the app involved. The
event aggregator receives the setting-change event, and it acts on it.

**Danger!** See change for a case where it's easy to call a method that needs a
read() on MidiSettings from a block that probably is holding a write() on it.
This was a bit harder to solve because of a silly issue where egui needs a
`Response`, and the normal `if let Ok(...)` code would leave a case where we
can't produce a `Response`, so it's easier to do a `write().unwrap()`, which
means the lock stays locked for a larger scope than we really want.