Work around a mysterious crash deep in egui
Switching to the placeholder track view mode made egui blow up, but only if
there was at least one SignalPath in the track. I can't repro it if I remove
that view mode. This is very unsatisfying, but the crash is so far from my own
code that it's really hard to tell what I'm doing to cause it, or where the bug
might be if it's in egui. So I'm just going to be happy I can't repro it
anymore.