Allow loading a drumkit as a group of samples (progress on #43)
Introduced KitLibrary. Both this and SampleLibrary are crude and fragile, but
they're capturing the dataflow that we need to build something better.

Have only a 707 library for now, and its mapping of notes to sounds is
nonstandard.

Next I'll introduce a way for an instrument in a track to provide a list of
notes it supports, which the composer can use to annotate the Y axis.