Experiment: make a Widget operate only on &, not &mut
I'm researching #37 - how to make the main loop more efficient, or at least
spread across more CPU cores.

I looked at [tower](https://docs.rs/tower/) because I'm thinking about an
async/await migration, which could make it a little easier to spread work across
a thread pool. That led to writing a little spike solution that runs egui in the
main thread, and has the engine in its own tokio::spawn area, doing an
async-based loop. This got me back to a problem I've been avoiding for a while:
if egui needs high-bandwidth, wide-scope access to Project's data, then it most
likely needs direct access to Project, which means that Project needs to be
behind at least an Arc<>, if not an Arc<Mutex> or Arc<RwLock>.

I could maintain a separate structure for egui that has all the project state
needed to draw it. Given the complexity of Project, I don't think that's a good
approach. Which is why I think egui needs direct access.

But this doesn't mean the whole Project needs to be behind a single mutex. It
could have several of them, maybe even many. To test this, I'm seeing how far I
can get with just Arc<Project>. The first roadblock is that egui widgets tend to
take &mut access to the things they want to modify. I use a hybrid of direct mut
access and Elm-style action-passing, so this CL makes ArrangementWidget fully
read-only, to get a sense of the work involved.

If I can get a good-sized area of the rendering code to work without a RwLock,
then I can see about re-granting write access but more fine-grained.