Fix #37: event loop not as busy
I almost checked in a complicated solution that didn't address the original
problem. Read on!

MiniDaw::spawn_app_channel_watcher() seems to have been the culprit. First, some
background.

As much as possible, we do work off the main/UI thread. The other threads that
do work communicate with the main thread using channels. The UI thread normally
gets time to run only when [App::update()] is called, and that normally happens
only when there's user input (mouse, keyboard) or someone has called
[Context::request_repaint()]. If we process UI channel messages only inside
update(), and update() isn't called for a long time, then UI messages could be
stale for a long time.

The original solution was to set up a loop in a separate thread
(`MiniDaw::spawn_app_channel_watcher()`) that looked to see whether the UI's
receive channel had anything in it. If it did, it requested a repaint.

It appears that this happened a *lot*, but there was more to it than that.

The first solution (to fix the original solution) was to kill
`spawn_app_channel_watcher()` entirely, and then try being smarter about when to
request a repaint. I focused on the visualization queue, which continuously
visualizes the audio output, and thus could have idled when the audio idled.
Adding that idle indeed quieted the busy loop, but I was a little sad because it
meant I'd now have to be very smart about when to redraw.

The second solution, which I'm checking in now, fixes the actual problem. This
is the original loop:

```rust
  let mut sel = Select::new();
  let _ = sel.recv(&receiver);
  loop {
      let _ = sel.ready();
      ctx.request_repaint();
  }
```

The problem was that `sel` remains in the ready state until the update() loop
actually reads off all the channel messages. Since update() runs relatively
rarely (60fps typically), and might be delayed until the next vsync, that could
be many milliseconds from the `ctx.request_repaint()`. So what happens? The loop
keeps looping! That was our real busy loop. In retrospect, I see how the
behavior fits better -- it was always pegging a single core near 100%, yet the
DAW itself was easily keeping up with the audio interface (no underruns), so
that meant it wasn't doing too much *real* work; it was doing infinite work.

The solution is to add a new channel from update() to
`spawn_app_channel_watcher()` telling it that the update cycle was done, and
that it could go look at `sel` again. This allows `spawn_app_channel_watcher()`
to sleep/block after the first `ready()` comes back and until update() is ready
to run again.

On my Ryzen 7 7735HS, the new code in a debug build settles around 3.9% CPU
total, with no single core above about 25%. Release is 1.3% total, max single
core 11%.

Previously, total CPU usage in debug was at least 9%, and the busy-looping
single core was around 77-80%.