Refactor many generate_next into generate()
I have redone this a few times, and I think it's going to stick this time
because I'm not requiring Generates implementers to maintain internal buffers.
Rather, the caller of Generates supplies it, but I made GenerationBuffer to make
that easier.

Rewrote FmSynth to do bulk generation. It's so-so. I don't know whether we'll
need a per-frame callback on generates() to let the caller modulate the
signal-generation parameters.